% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbors cross-validation predicting function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame used as covariates.}

\item{cl}{True class value of the training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list with objects : class, a vector of the predicted class Ŷ i for
all observations, and cv_err, a numeric with the cross-validation
misclassification error.
}
\description{
This function builds a k-nearest neighbors cross-validation predicting model.
}
\examples{
my_knn_cv(data, cl, 5, 5)
my_knn_cv(data, cl, 1, 10)

}
