% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object : a symbolic description of the model
to be fitted.}

\item{data}{Input data frame.}
}
\value{
A coefficient table with rows for each coefficient (including the
(Intercept)!) and columns for the Estimate, Std. Error, t value, and
Pr(>|t|).
}
\description{
This function fits a linear model.
}
\examples{
my_lm(y ~ x ^ 2, dataset)
my_lm(y ~ x, dataset)

}
\keyword{inference}
