% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random forest cross-validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Integer representing the number of folds.}
}
\value{
A numeric with the cross-validation error.
}
\description{
This function will predict body_mass_g by building and using a random forest
cross-validation model, which uses covariates bill_length_mm,
bill_depth_mm, and flipper_length_mm.
}
\examples{
my_rf_cv(5)
my_rf_cv(10)

}
\keyword{prediction}
