% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T_test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data used to find its t-test results.}

\item{alternative}{A character string specifying the alternative hypothesis,
and it only accepts "two.sided", "less", or "greater".}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements test_stat, the numeric test statistic, df, the
degrees of freedom, \code{alternative} and p_val, the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\keyword{inference}
